+++ src/test/kotlin/org/jetbrains/intellij/tasks/BuildSearchableOptionsTaskSpec.kt
@@ -1,9 +1,7 @@
 package org.jetbrains.intellij.tasks
 
-import org.gradle.util.VersionNumber
 import org.jetbrains.intellij.IntelliJPluginConstants
 import org.jetbrains.intellij.SearchableOptionsSpecBase
-import org.junit.Assume.assumeFalse
 import kotlin.test.Test
 import kotlin.test.assertTrue
 
@@ -45,8 +43,6 @@ class BuildSearchableOptionsTaskSpec : SearchableOptionsSpecBase() {
 
     @Test
     fun `reuse configuration cache`() {
-        assumeFalse("Feature supported since Gradle 6.6", VersionNumber.parse(gradleVersion) < VersionNumber.parse("6.6"))
-
         build(IntelliJPluginConstants.BUILD_SEARCHABLE_OPTIONS_TASK_NAME, "--configuration-cache")
         val result = build(IntelliJPluginConstants.BUILD_SEARCHABLE_OPTIONS_TASK_NAME, "--configuration-cache")