+++ src/test/kotlin/org/jetbrains/intellij/tasks/PatchPluginXmlTaskSpec.kt
@@ -380,8 +380,6 @@ class PatchPluginXmlTaskSpec : IntelliJPluginSpecBase() {
 
     @Test
     fun `reuse configuration cache`() {
-        assumeFalse("Feature supported since Gradle 6.6", VersionNumber.parse(gradleVersion) < VersionNumber.parse("6.6"))
-
         pluginXml.xml("""
             <idea-plugin />
         """)