+++ src/test/kotlin/org/jetbrains/intellij/tasks/PublishPluginTaskSpec.kt
@@ -1,9 +1,7 @@
 package org.jetbrains.intellij.tasks
 
-import org.gradle.util.VersionNumber
 import org.jetbrains.intellij.IntelliJPluginConstants
 import org.jetbrains.intellij.IntelliJPluginSpecBase
-import org.junit.Assume.assumeFalse
 import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertTrue
@@ -40,8 +38,6 @@ class PublishPluginTaskSpec : IntelliJPluginSpecBase() {
 
     @Test
     fun `reuse configuration cache`() {
-        assumeFalse("Feature supported since Gradle 6.6", VersionNumber.parse(gradleVersion) < VersionNumber.parse("6.6"))
-
         buildFile.groovy("""
             publishPlugin { }
             verifyPlugin {