+++ src/test/kotlin/org/jetbrains/intellij/tasks/RunPluginVerifierTaskSpec.kt
@@ -251,8 +251,6 @@ class RunPluginVerifierTaskSpec : IntelliJPluginSpecBase() {
 
     @Test
     fun `reuse configuration cache`() {
-        assumeFalse("Feature supported since Gradle 6.6", VersionNumber.parse(gradleVersion) < VersionNumber.parse("6.6"))
-
         writePluginXmlFile()
         buildFile.groovy("""
             version = "1.0.0"