+++ src/test/kotlin/org/jetbrains/intellij/tasks/SignPluginTaskSpec.kt
@@ -1,9 +1,7 @@
 package org.jetbrains.intellij.tasks
 
-import org.gradle.util.VersionNumber
 import org.jetbrains.intellij.IntelliJPluginConstants
 import org.jetbrains.intellij.IntelliJPluginSpecBase
-import org.junit.Assume.assumeFalse
 import kotlin.test.Test
 import kotlin.test.assertTrue
 
@@ -11,8 +9,6 @@ class SignPluginTaskSpec : IntelliJPluginSpecBase() {
 
     @Test
     fun `reuse configuration cache`() {
-        assumeFalse("Feature supported since Gradle 6.6", VersionNumber.parse(gradleVersion) < VersionNumber.parse("6.6"))
-
         build(IntelliJPluginConstants.SIGN_PLUGIN_TASK_NAME, "--configuration-cache")
         val result = build(IntelliJPluginConstants.SIGN_PLUGIN_TASK_NAME, "--configuration-cache")